requirements.txt
pdfplumber
pymupdf
pytesseract
opencv-python
pillow
sentence-transformers
transformers
torch
torchvision
chromadb
scikit-learn
openai
streamlit
